Revert main.rb name back to app.rb and make an entry point for the app on main.rb
